+++ CHANGELOG.md
@@ -8,6 +8,33 @@ This project uses [*towncrier*](https://towncrier.readthedocs.io/) and the chang
 
 <!-- towncrier release notes start -->
 
+## [3.197.1](https://github.com/metalbear-co/mirrord/tree/3.197.1) - 2026-03-31
+
+
+### Fixed
+
+- Fixed windows python `socketpair` regression. affecting versions
+  3.190.0-3.196.0
+
+
+### Internal
+
+- RabbitMQ missing field from ConnectParams.
+- Fixed `outgoing_traffic_filter_udp_with_connect` e2e test.
+
+## [3.197.0](https://github.com/metalbear-co/mirrord/tree/3.197.0) - 2026-03-31
+
+
+### Added
+
+- Added RabbitMQ queue splitting support.
+
+
+### Changed
+
+- Check for existing key (preview env) and warn the user.
+- `mirrord ci start --foreground` command behaves exactly like `mirrord exec`.
+
 ## [3.196.0](https://github.com/metalbear-co/mirrord/tree/3.196.0) - 2026-03-28